Fixup_bundle handles Qt5 and has better argument handling
The fixup_bundle.py script was not handling its arguments correctly and
failing for Qt5's frameworks since there were multiple files with the
same name in the Qt5 directory.  This also modifies the argument
handling of fixup_bundle so that arguments can be more easily passed in
and modified to add new search paths.