windows: Make Qt5 part of the tomviz component
For some reason this is needed for the NSIS generator to install the libs.
The ZIP generator worked fine without it, which is what I was testing.
There is still an issue with Qt though, starting the application installed
from the NSIS installer says Qt cannot find the 'windows' plugin.