linux: install qt plugins and their dependencies
This is actually several interdependent pieces.  First the plugins are
installed to the plugins directory under the install location.  They
have to be there since Qt assumes that ../lib from the plugins directory
is where the libraries are.  Then it installs a qt.conf in bin telling
Qt that its libraries live in ../plugins.  Finally each plugin must be
parsed for dependencies that need to be installed to make sure the
plugins don't have unsatisfied dependencies which will cause tomviz to
crash when trying to load.