Update dependency installation rules for linux
With CMake 3.5 (required by newer paraview) the get_prerequisites
function changed to error out and quit rather than return an incomplete
list of dependencies.  So we need to pass in all the locations that
dependencies are located so that it can find them all and return a list
rather than failing.

Then I had to update the logic so that it didn't install unnecessary
duplicates of libraries in the lib/paraview folder.  I made it search
for the library in the new package and only install it if it was not
found.