Disable strict path checking in tomviz configure step
Tomviz by default checks if ParaView and ITK are compiled against the
same python.  Since the superbuild uses a pre-built ITK, the paths for
ITK will be different from the path to python in the superbuild and this
will fail.  This commit disables these checks for the superbuild as the
bundled ITK causes false positives on these checks and makes the build
fail.