Exclude vcruntime*.dll from win32 python install
If vcruntime*.dll is included in the python install, then two
copies of it get installed to the same location: one from the
python install, and one from just installing the required standard
libraries.

Remove it from the python install list so it is not included. This
fixes an error with WIX.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>