Don't install ITK and Tomviz libs twice on Windows
The entire `lib` directory is already installed, so we should hopefully
not need to install the `lib/tomviz` and `lib/itk` directories again.

This will hopefully fix some of the windows package bloating size.
But we should probably check the Windows package and make sure ITK
still works after this gets merged and built.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>