On Linux, package *libcrypt* as a dependency
Some operating systems such as fedora do not have the same version
of libcrypt. We should package libcrypt so that it can be used.

This packaging method works by overriding the type of the libcrypt file
as it is seen by `get_prerequisites()` so that it will not be ignored.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>