Do not specify install prefix for pygments
Specifying the install prefix breaks the windows build, with it
complaining that the install prefix is not in the python path.

However, it appears to install it to the correct location without
the install prefix. Linux also appears to install it to the
correct location without the install prefix. Just need to make sure
it installs to the correct location on mac, then...

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>