Install pygments using pip
With "python setup.py install", the installed pygments directory
in "site-packages" has a version number and a ".egg" at the end.
It relies on the use of easypath to be found.

If we install it with "pip install ." instead, the installed
pygments directory is just named "pygments", and can be found
without easypath.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>