Add option to install python project with pip
Many projects are more easily locatable by tomviz when they are
installed with pip (because otherwise, they may have a versioned
"*.egg" file and rely on easypath). However, some projects cannot
be installed with pip (pkgconfig, in particular).

Add an option to install the python project with pip, and
use the option for a few projects where it is needed.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>