Fix Mac and Windows Qt plugins
Mac was producing an error when trying to fixup some of the plugins
it was trying to install. Specifically, it produced an error for a
debug library of libqsqlite. Thus, we should specify all the plugins
we want directly to avoid these errors.

For Windows, we need to find QSvg to use the targets.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>